Move MockConverter to separate file
