+++ src/Converters/Converters.Tests/Converters.Tests.csproj
@@ -56,6 +56,7 @@
     <Compile Include="HexToColorTests.cs" />
     <Compile Include="InvertedBooleanTests.cs" />
     <Compile Include="LowerTextTests.cs" />
+    <Compile Include="Mocks\MockConverter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UpperTextTests.cs" />
     <Compile Include="ValueConverterGroupTests.cs" />
